+++ dd-java-agent/instrumentation/instrumentation.gradle
@@ -9,35 +9,31 @@ plugins {
 }
 apply from: "${rootDir}/gradle/java.gradle"
 
-// add all subprojects under 'instrumentation' to the agent's dependencies
-def skip = ["connection-error-handling"]
 Project instr_project = project
 subprojects { subProj ->
-  if (!skip.contains(name)) {
-    apply plugin: "net.bytebuddy.byte-buddy"
-    apply plugin: 'muzzle'
+  apply plugin: "net.bytebuddy.byte-buddy"
+  apply plugin: 'muzzle'
 
+  subProj.byteBuddy {
+    transformation {
+      // Applying NoOp optimizes build by applying bytebuddy plugin to only compileJava task
+      tasks = ['compileJava', 'compileScala']
+      plugin = 'datadog.trace.agent.tooling.muzzle.MuzzleGradlePlugin$NoOp'
+    }
+  }
+
+  subProj.afterEvaluate {
     subProj.byteBuddy {
       transformation {
-        // Applying NoOp optimizes build by applying bytebuddy plugin to only compileJava task
         tasks = ['compileJava', 'compileScala']
-        plugin = 'datadog.trace.agent.tooling.muzzle.MuzzleGradlePlugin$NoOp'
+        plugin = 'datadog.trace.agent.tooling.muzzle.MuzzleGradlePlugin'
+        classPath = project(':dd-java-agent:agent-tooling').configurations.instrumentationMuzzle + subProj.configurations.compile + subProj.sourceSets.main.output
       }
     }
+  }
 
-    subProj.afterEvaluate {
-      subProj.byteBuddy {
-        transformation {
-          tasks = ['compileJava', 'compileScala']
-          plugin = 'datadog.trace.agent.tooling.muzzle.MuzzleGradlePlugin'
-          classPath = project(':dd-java-agent:agent-tooling').configurations.instrumentationMuzzle + subProj.configurations.compile + subProj.sourceSets.main.output
-        }
-      }
-    }
-
-    instr_project.dependencies {
-      compile(project(subProj.getPath()))
-    }
+  instr_project.dependencies {
+    compile(project(subProj.getPath()))
   }
 }
