+++ dd-java-agent/instrumentation/jax-rs-client/connection-error-handling-jersey/connection-error-handling-jersey.gradle
@@ -17,4 +17,6 @@ dependencies {
   implementation deps.autoservice
 
   compile project(':dd-java-agent:agent-tooling')
+
+  compileOnly project(':dd-java-agent:instrumentation:jax-rs-client')
 }
