+++ dd-java-agent/instrumentation/jax-rs-client/connection-error-handling-resteasy/connection-error-handling-resteasy.gradle
@@ -17,4 +17,6 @@ dependencies {
   implementation deps.autoservice
 
   compile project(':dd-java-agent:agent-tooling')
+
+  compileOnly project(':dd-java-agent:instrumentation:jax-rs-client')
 }
