+++ dd-java-agent/instrumentation/jax-rs-client/connection-error-handling-resteasy/src/main/java/datadog/trace/instrumentation/connection_error/resteasy/ResteasyClientConnectionErrorInstrumentation.java
@@ -1,5 +1,6 @@
-package datadog.trace.instrumentation.connection_error.jersey;
+package datadog.trace.instrumentation.connection_error.resteasy;
 
+import static io.opentracing.log.Fields.ERROR_OBJECT;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
@@ -7,6 +8,7 @@ import static net.bytebuddy.matcher.ElementMatchers.returns;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.instrumentation.jaxrs.ClientTracingFilter;
 import io.opentracing.Span;
 import io.opentracing.log.Fields;
 import io.opentracing.tag.Tags;
@@ -20,22 +22,22 @@ import java.util.concurrent.TimeoutException;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.glassfish.jersey.client.ClientRequest;
+import org.jboss.resteasy.client.jaxrs.internal.ClientConfiguration;
 
 /**
  * JAX-RS Client API doesn't define a good point where we can handle connection failures, so we must
  * handle these errors at the implementation level.
  */
 @AutoService(Instrumenter.class)
-public final class JerseyClientConnectionErrorInstrumentation extends Instrumenter.Default {
+public final class ResteasyClientConnectionErrorInstrumentation extends Instrumenter.Default {
 
-  public JerseyClientConnectionErrorInstrumentation() {
+  public ResteasyClientConnectionErrorInstrumentation() {
     super("jax-rs", "jaxrs", "jax-rs-client");
   }
 
   @Override
   public ElementMatcher<TypeDescription> typeMatcher() {
-    return named("org.glassfish.jersey.client.JerseyInvocation");
+    return named("org.jboss.resteasy.client.jaxrs.internal.ClientInvocation");
   }
 
   @Override
@@ -57,14 +59,14 @@ public final class JerseyClientConnectionErrorInstrumentation extends Instrument
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
     public static void handleError(
-        @Advice.FieldValue("requestContext") final ClientRequest context,
+        @Advice.FieldValue("configuration") final ClientConfiguration context,
         @Advice.Thrown final Throwable throwable) {
       if (throwable != null) {
-        final Object prop = context.getProperty(WrappedFuture.SPAN_PROPERTY_NAME);
+        final Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
         if (prop instanceof Span) {
           final Span span = (Span) prop;
           Tags.ERROR.set(span, true);
-          span.log(Collections.singletonMap(Fields.ERROR_OBJECT, throwable));
+          span.log(Collections.singletonMap(ERROR_OBJECT, throwable));
           span.finish();
         }
       }
@@ -75,19 +77,20 @@ public final class JerseyClientConnectionErrorInstrumentation extends Instrument
 
     @Advice.OnMethodExit(suppress = Throwable.class)
     public static void handleError(
-        @Advice.FieldValue("requestContext") final ClientRequest context,
+        @Advice.FieldValue("configuration") final ClientConfiguration context,
         @Advice.Return(readOnly = false) Future future) {
-      future = new WrappedFuture(future, context);
+      if (!(future instanceof WrappedFuture)) {
+        future = new WrappedFuture(future, context);
+      }
     }
   }
 
   public static class WrappedFuture<T> implements Future<T> {
-    public static final String SPAN_PROPERTY_NAME = "datadog.trace.jaxrs.span"; // Copied elsewhere
 
     private final Future<T> wrapped;
-    private final ClientRequest context;
+    private final ClientConfiguration context;
 
-    public WrappedFuture(final Future<T> wrapped, final ClientRequest context) {
+    public WrappedFuture(final Future<T> wrapped, final ClientConfiguration context) {
       this.wrapped = wrapped;
       this.context = context;
     }
@@ -112,7 +115,7 @@ public final class JerseyClientConnectionErrorInstrumentation extends Instrument
       try {
         return wrapped.get();
       } catch (final ExecutionException e) {
-        final Object prop = context.getProperty(SPAN_PROPERTY_NAME);
+        final Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
         if (prop instanceof Span) {
           final Span span = (Span) prop;
           Tags.ERROR.set(span, true);
@@ -129,7 +132,7 @@ public final class JerseyClientConnectionErrorInstrumentation extends Instrument
       try {
         return wrapped.get(timeout, unit);
       } catch (final ExecutionException e) {
-        final Object prop = context.getProperty(SPAN_PROPERTY_NAME);
+        final Object prop = context.getProperty(ClientTracingFilter.SPAN_PROPERTY_NAME);
         if (prop instanceof Span) {
           final Span span = (Span) prop;
           Tags.ERROR.set(span, true);
