+++ dd-java-agent/instrumentation/jax-rs-client/jax-rs-client.gradle
@@ -29,8 +29,8 @@ dependencies {
 
   testCompile project(':dd-java-agent:testing')
 
-  testCompile project(':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling:jersey')
-  testCompile project(':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling:resteasy')
+  testCompile project(':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling-jersey')
+  testCompile project(':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling-resteasy')
 
   testCompile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0.1'
