+++ settings.gradle
@@ -25,8 +25,8 @@ include ':dd-java-agent:instrumentation:http-url-connection'
 include ':dd-java-agent:instrumentation:hystrix-1.4'
 include ':dd-java-agent:instrumentation:jax-rs-annotations'
 include ':dd-java-agent:instrumentation:jax-rs-client'
-include ':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling:jersey'
-include ':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling:resteasy'
+include ':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling-jersey'
+include ':dd-java-agent:instrumentation:jax-rs-client:connection-error-handling-resteasy'
 include ':dd-java-agent:instrumentation:java-concurrent'
 include ':dd-java-agent:instrumentation:java-concurrent:scala-testing'
 include ':dd-java-agent:instrumentation:java-concurrent:akka-testing'
